% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_palette.R
\name{af_palette}
\alias{af_palette}
\title{Analysis Function (AF) palette}
\usage{
af_palette(...)
}
\arguments{
\item{...}{Integer or colour name(s) as character strings.
Eight colours are available:
\itemize{
\item dark_blue
\item turquoise
\item dark_pink
\item orange
\item dark_grey
\item light_purple
\item red
\item gold
}}
}
\value{
Returns a named character vector.
}
\description{
Returns name and hex code for Government Analysis Function (AF) colours.
Calling the function with no arguments returns all colours.
If an integer (n) is passed to the function the nth colour is returned.
If the name(s) of specific colours are passed to the function, the name and hex
code of those colours are returned.
}
\details{
See \link[AF data visualisation colour guidance]{https://analysisfunction.civilservice.gov.uk/policy-store/data-visualisation-colours-in-charts/}
}
\examples{
af_palette(1)
}
