% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_palette.R
\name{fc_palette}
\alias{fc_palette}
\title{Forestry Commission (FC) palette}
\usage{
fc_palette(...)
}
\arguments{
\item{...}{Integer or colour name(s) as character strings.
Two colours are available:
\itemize{
\item broadleaf_green
\item conifer_green
}}
}
\value{
Returns a named character vector.
}
\description{
Returns name and hex code for the Forestry Commission's (FC) colours.
Calling the function with no arguments returns both colours.
If an integer (n) is passed to the function the nth colour is returned.
If the name(s) of specific colours are passed to the function, the name and hex code of those colours are returned.

There are two colours in the FC palette:
\itemize{
\item broadleaf_green
\item conifer_green
}
}
\details{
See \link[Forestry Commission]{https://www.gov.uk/government/organisations/forestry-commission} for the colours in action.
}
\note{
I no longer work at Forest Research, the research agency of the Forestry Commission.
}
\examples{
fc_palette(1)
}
\author{
Si Maxwell \email{simon.maxwell@ukhsa.gov.uk}
}
