% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggGSS-package.R
\docType{package}
\name{ggGSS-package}
\alias{ggGSS}
\alias{ggGSS-package}
\title{ggGSS: Produce data visualisations formatted to UK Government Analysis Function best practice guidance}
\description{
An R pacakge to produce data visualisations formatted in compilance with best practice guidance from the UK Government's Analysis Function.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/simaxwell/ggGSS}
  \item Report bugs at \url{https://github.com/simaxwell/ggGSS/issues}
}

}
\keyword{internal}
