% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_palette.R
\name{scale_colour_af_d}
\alias{scale_colour_af_d}
\title{AF-style scale functions (colour) for ggplot2}
\usage{
scale_colour_af_d(palette = "af", direction = 1, ...)
}
\arguments{
\item{palette}{The palette name: either AF (default) or FC.}

\item{direction}{For FC palette only: sets the order of the colours. If 1 (the default) colours are as output by fc_palette. If -1, the order of colours is reversed.}

\item{...}{Parameters to \code{\link[ggplot2]{discrete_scale}}.}
}
\value{
Returns a discrete ggplot2 scale function.
}
\description{
Create a discrete ggplot2 colour scale function for the AF colour palette.
}
\examples{
ggplot2::discrete_scale("colour", "af", generate_palette("af"))
}
\author{
Si Maxwell \email{simon.maxwell@ukhsa.gov.uk}
}
