% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_palette.R
\name{scale_fill_gss_d}
\alias{scale_fill_gss_d}
\title{GSS-style scale functions (fill) for ggplot2}
\usage{
scale_fill_gss_d(palette = "gss", direction = 1, ...)
}
\arguments{
\item{palette}{The palette name: either GSS (default) or FC.}

\item{direction}{Sets the order of colours. If 1 (the default) colours are as output by fc_palette. If -1, the order of colours is reversed. For FC palette only.}

\item{...}{Parameters to \code{\link[ggplot2]{discrete_scale}}.}
}
\value{
Returns a ggplot2 scale function.
}
\description{
Create a discrete ggplot2 fill scale function for the GSS colour palette.
}
\examples{
ggplot2::discrete_scale("fill", "gss", generate_palette("gss"))
}
\author{
Si Maxwell \email{simon.maxwell@forestresearch.gov.uk}
}
