% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukhsa_palette.R
\name{ukhsa_palette}
\alias{ukhsa_palette}
\title{UK Health Security Agency (UKHSA) palette}
\usage{
ukhsa_palette(...)
}
\arguments{
\item{...}{Integer or colour name(s) as character strings.
Six colours are available:
\itemize{
\item teal
\item midnight
\item plum
\item moonlight
\item wine
\item cherry
}}
}
\value{
Returns a named character vector.
}
\description{
Returns name and hex code for UK Health Security Agency (UKHSA) colours.
Calling the function with no arguments returns all colours.
If an integer (n) is passed to the function the nth colour is returned.
If the name(s) of specific colours are passed to the function, the name and hex code of those colours are returned.
}
\details{
See UKHSA brand guidance.
}
\examples{
ukhsa_palette(1)
}
\author{
Si Maxwell \email{simon.maxwell@ukhsa.gov.uk}
}
