% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_selector.R
\name{palette_selector}
\alias{palette_selector}
\title{Palette selector}
\usage{
palette_selector(palette = "GSS")
}
\arguments{
\item{palette}{A character string indicating the colour palette to use:
\itemize{
\item "GSS" (or "gss")
\item "FC" (or "fc")
\item "UKHSA" (or "ukhsa")
}}
}
\value{
Returns a named character vector.
}
\description{
Select the palette of interest: GSS (default) or FC
}
\examples{
palette_selector(palette = "GSS")
}
\author{
Si Maxwell \email{simon.maxwell@ukhsa.gov.uk}
}
